Change Pod.formats property to Pod.format (dict)
Property Pod.format returns a dictionary of available formats mapping
to a list of the all the subpod's contents of that format. For example,
pod.format['plaintext'] returns a list of all the plaintext subpod text.

Image requires special handling, and is packaged as another dictionary,
with url, alt, title, width, height as keys.